The 'inspect' command was extracted into a separate file.
