+++ src/cmd_god/stat.cpp
@@ -1,14 +1,11 @@
 #include "stat.h"
 
 #include "ban.h"
-#include "entities/char_data.h"
 #include "entities/char_player.h"
 #include "entities/player_races.h"
 #include "utils/utils_char_obj.inl"
 #include "description.h"
 #include "game_fight/fight_hit.h"
-#include "game_fight/pk.h"
-#include "handler.h"
 #include "olc/olc.h"
 #include "game_mechanics/glory.h"
 #include "game_mechanics/glory_const.h"
