+++ src/game_mechanics/guilds.cpp
@@ -103,13 +103,11 @@ const std::string &GuildInfo::GetMsg(EMsg msg_id) {
 		{EMsg::kAskToRoom, "$n попросил$u в ученики к $N2."},
 		{EMsg::kLearnToChar, "Вы получили несколько уроков и мудрых советов от $N1."},
 		{EMsg::kLearnToRoom, "$N дал$G $n2 несколько наставлений."},
-		{EMsg::kAllSkills,
-		 "$N сказал$G: '$n, нельзя научиться всем умениям или способностям сразу. Выбери необходимые!'"},
+		{EMsg::kAllSkills, "$N сказал$G: '$n, нельзя научиться всем умениям или способностям сразу. Выбери необходимые!'"},
 		{EMsg::kTalentEarned, "Под наставничеством $N1 вы изучили {} {}'{}'{}."},
 		{EMsg::kNothingLearned, "$N ничему новому вас не научил$G."},
 		{EMsg::kListEmpty, "$N сказал$G : 'Похоже, $n, я не смогу тебе помочь'."},
-		{EMsg::kIsInsolvent,
-		 "$N сказал$G : 'Вот у меня забот нет - голодранцев наставлять! Иди-ка, $n, подзаработай сначала!"},
+		{EMsg::kIsInsolvent, "$N сказал$G : 'Вот у меня забот нет - голодранцев наставлять! Иди-ка, $n, подзаработай сначала!"},
 		{EMsg::kFree, "бесплатно"},
 		{EMsg::kTemporary, "временно"},
 		{EMsg::kYouGiveMoney, "Вы дали {} $N2."},
