+++ src/structs/global_objects.h
@@ -13,6 +13,7 @@
 #include "heartbeat.h"
 #include "speedwalks.h"
 #include "cmd_god/shutdown_parameters.h"
+#include "cmd_god/inspect.h"
 #include "game_economics/shops_implementation.h"
 #include "entities/world_objects.h"
 #include "entities/world_characters.h"
